Add `uintptr_t` to main.c and `sys_cfb.c`

+++ include/libc/stddef.h
@@ -3,9 +3,9 @@
 
 #define NULL ((void*)0)
 
-typedef unsigned long size_t;
+typedef unsigned int size_t;
 
-typedef unsigned int uintptr_t;
+typedef signed int ptrdiff_t;
 
 #ifdef __GNUC__
 #define offsetof(structure, member) __builtin_offsetof (structure, member)
